1. moved to NUnit 2.4.2 2. enabled shared memory testing for all platforms in Client.build 3. Added some casts to fix float comparison using nunit 2.4.2

+++ TestSuite/Source/CultureTests.cs
@@ -80,7 +80,7 @@ namespace MySql.Data.MySqlClient.Tests
 				if (prepared) cmd.Prepare();
 				reader = cmd.ExecuteReader();
 				reader.Read();
-				Assert.AreEqual(2.3, reader.GetFloat(0));
+				Assert.AreEqual(2.3, (decimal)reader.GetFloat(0));
 				Assert.AreEqual(4.6, reader.GetDouble(1));
 				Assert.AreEqual(23.82, reader.GetDecimal(2));
 			}
